+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Remove the `UMap` module and the `umap` benchmarks cabal target.
 * Export `dRepToText`
+* Deprecated `bheader` and `bbody`
+* Add field accessors to `Block`: `blockHeader` and `blockBody`.
+* Expose `dRepToText`
 * Modify `withdrawalsThatDoNotDrainAccounts` to return `Maybe (Withdrawals, Withdrawals)` where the `fst` are either missing accounts or in the wrong network and `snd` are incomplete withdrawals.
 * Add `FromJSON` instance for `PParamUpdate`